fix: prey

+++ src/enums.h
@@ -159,14 +159,14 @@ enum SpellGroup_t : uint8_t {
 };
 
 // New Prey
-enum PreySlotNum_t : uint8_t
+enum PreySlotNum_t : uint16_t
 {
 	PREY_SLOTNUM_FIRST = 0,
 	PREY_SLOTNUM_SECOND = 1,
 	PREY_SLOTNUM_THIRD = 2,
 };
 
-enum PreySlotStatus_t : uint8_t
+enum PreySlotStatus_t : uint16_t
 {
 	PREY_SLOT_LOCKED = 0,
 	PREY_SLOT_UNLOCKED = 1,
